refactor(foundation): Move common utility classes to foundation-common module
- Move ConditionWaiter and MuteExceptionUtil classes from service-registry and dynamic-config modules to foundation-common module - Update import paths in related modules - Optimize EtcdConfigIT test class, add retry mechanism and log output
