+++ foundations/foundation-common/src/main/java/org/apache/servicecomb/foundation/common/utils/ConditionWaiter.java
@@ -14,7 +14,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.apache.servicecomb.registry.etcd;
+package org.apache.servicecomb.foundation.common.utils;
 
 import java.util.concurrent.Callable;
 import java.util.concurrent.CompletableFuture;
